refactor: Refactor encoder layer in PerStreamPBE
The encoder layer in PerStreamPBE is refactored to use
TransformerEncoderLayer instead of a custom EncoderLayer. This
change aims to leverage the standard TransformerEncoderLayer and
simplify the encoder structure. The self-attention mechanism is
explicitly assigned to the encoder layer. Also, the number of
communication layers in SiFormer is increased from 1 to 3.